+++ packages/tests/src/tests/menubar/menubar-test.svelte
@@ -6,10 +6,12 @@
 </script>
 
 <main>
+	<button data-testid="previous-button">previous button</button>
 	<Menubar.Root {...restProps} data-testid="root">
 		<MenubarMenu id="1" />
 		<MenubarMenu id="2" />
 		<MenubarMenu id="3" />
 		<MenubarMenu id="4" />
 	</Menubar.Root>
+	<button data-testid="next-button">next button</button>
 </main>